+++ tests/v1_test.py
@@ -899,14 +899,15 @@ def test_deprecated_expr_methods() -> None:
     pytest.importorskip("polars")
     import polars as pl
 
-    data = {"a": [0, 0, 2, -1], "b": [1, 3, 2, None]}
+    data = {"a": [0, 0, 2, -1]}
     df = nw_v1.from_native(pl.DataFrame(data), eager_only=True)
     result = df.select(
         c=nw_v1.col("a").sort().head(2),
         d=nw_v1.col("a").sort().tail(2),
         e=(nw_v1.col("a") == 0).arg_true(),
+        f=nw_v1.col("a").gather_every(2),
     )
-    expected = {"c": [-1, 0], "d": [0, 2], "e": [0, 1]}
+    expected = {"c": [-1, 0], "d": [0, 2], "e": [0, 1], "f": [0, 2]}
     assert_equal_data(result, expected)
 
     with pytest.deprecated_call():
@@ -914,4 +915,5 @@ def test_deprecated_expr_methods() -> None:
             c=nw.col("a").sort().head(2),
             d=nw.col("a").sort().tail(2),
             e=(nw.col("a") == 0).arg_true(),
+            f=nw.col("a").gather_every(2),
         )
